requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
huggingface_hub
pdfplumber
PyMuPDF
faiss-cpu
scikit-learn
sentence-transformers
